🏗️ Architect: Modularized catalogs
- Refactored `apts/catalogs.py` into a modular package `apts/catalogs/`.
- Decomposed catalog loading logic into `messier.py`, `ngc.py`, and `stars.py`.
- Moved `Catalogs` class and `initialize_catalogs` to `base.py`.
- Maintained public API via `__init__.py`.
- Updated internal references and verified with tests.
