--- apts/catalogs.py
@@ -1,296 +1,0 @@
-import logging
-import urllib.parse
-from importlib import resources
-from typing import cast
-
-import pandas as pd
-from skyfield.api import Star
-
-from .constants.constellations import constellation_map
-from .constants.objecttablelabels import ObjectTableLabels
-from .constants.strategies import DSOType
-from .units import get_unit_registry
-
-_messier_df = None
-_bright_stars_df = None
-_ngc_df = None
-
-logger = logging.getLogger(__name__)
-
-
-def _load_messier_with_units():
-    # Load Messier catalogue data
-    messier_df = pd.read_csv(str(resources.files("apts").joinpath("data/messier.csv")))
-
-    # Set proper dtypes for string columns
-    string_columns = ["Messier", "NGC", "Name", "Type", "Constellation"]
-    for column in string_columns:
-        messier_df[column] = messier_df[column].astype("string")
-
-    # Map Messier types to DSOType
-    messier_type_map = {
-        "Diffuse Nebula": DSOType.DN,
-        "Double Star": DSOType.DS,
-        "Elliptical Galaxy": DSOType.GX,
-        "Globular Cluster": DSOType.GC,
-        "Group/Asterism": DSOType.AST,
-        "Irregular Galaxy": DSOType.GX,
-        "Lenticular (S0) Galaxy": DSOType.GX,
-        "Open Cluster": DSOType.OC,
-        "Planetary Nebula": DSOType.PN,
-        "Spiral Galaxy": DSOType.GX,
-        "Star Cloud": DSOType.SC,
-        "Supernova Remnant": DSOType.SNR,
-        'Pleiades. Subaru."': DSOType.OC,  # Special case for M45 entry
-    }
-    messier_df[ObjectTableLabels.DSO_TYPE] = (
-        messier_df["Type"]
-        .map(messier_type_map.get)
-        .fillna(cast(DSOType, DSOType.OTHER))
-    )
-
-    # Pre-calculate Skyfield objects using raw floats before wrapping in Quantities
-    # Optimization: avoiding row-wise apply and costly Quantity.to().magnitude calls
-    messier_df["skyfield_object"] = [  # type: ignore
-        Star(ra_hours=ra, dec_degrees=dec)
-        for ra, dec in zip(messier_df["RA"], messier_df["Dec"])
-    ]
-
-    # Store float versions for performance-critical filtering and calculations
-    # to avoid Pint and Skyfield object overhead in high-frequency loops.
-    messier_df["ra_hours"] = messier_df["RA"].values
-    messier_df["dec_degrees"] = messier_df["Dec"].values
-    messier_df["Magnitude_float"] = messier_df["Magnitude"].values
-
-    # Convert columns to quantities with units (vectorized)
-    # Optimization: list(values * unit) is much faster than Series.apply()
-    ureg = get_unit_registry()
-    messier_df["Distance"] = list(messier_df["Distance"].values * ureg.light_year)
-    messier_df["RA"] = list(messier_df["RA"].values * ureg.hour)
-    messier_df["Dec"] = list(messier_df["Dec"].values * ureg.degree)
-    messier_df[ObjectTableLabels.SIZE_MAJOR] = list(
-        messier_df["Width"].values * ureg.arcminute
-    )
-    messier_df[ObjectTableLabels.SIZE_MINOR] = list(
-        messier_df["Height"].values * ureg.arcminute
-    )
-    messier_df["Magnitude"] = list(messier_df["Magnitude"].values * ureg.mag)
-
-    # Drop redundant columns
-    messier_df.drop(columns=["Width", "Height"], inplace=True)
-
-    # Add external links (vectorized list comprehension)
-    quoted_messier = [urllib.parse.quote(str(x)) for x in messier_df["Messier"]]
-    messier_df[ObjectTableLabels.SIMBAD] = [
-        f"https://simbad.u-strasbg.fr/simbad/sim-basic?Ident={q}"
-        for q in quoted_messier
-    ]
-    messier_df[ObjectTableLabels.ALADIN] = [
-        f"https://aladin.cds.unistra.fr/AladinLite/?target={q}" for q in quoted_messier
-    ]
-    messier_df[ObjectTableLabels.ASTROBIN] = [
-        f"https://www.astrobin.com/search/?q={q}" for q in quoted_messier
-    ]
-
-    return messier_df
-
-
-def _load_ngc_with_units():
-    # Load NGC catalogue data
-    ngc_df = pd.read_csv(
-        str(resources.files("apts").joinpath("data/ngc.csv")), sep=";", na_values=[""]
-    )
-
-    # Rename columns for consistency
-    ngc_df.rename(
-        columns={"Const": "Constellation", "V-Mag": "Magnitude"},
-        inplace=True,
-    )
-
-    # Mapping for NGC types to DSOType
-    ngc_type_map = {
-        "*": DSOType.OTHER,
-        "**": DSOType.DS,
-        "*Ass": DSOType.AST,
-        "Cl+N": DSOType.OC,  # Cluster with nebulosity
-        "Dup": DSOType.OTHER,
-        "EmN": DSOType.EN,
-        "G": DSOType.GX,
-        "GCl": DSOType.GC,
-        "GGroup": DSOType.AST,
-        "GPair": DSOType.GX,
-        "GTrpl": DSOType.GX,
-        "HII": DSOType.EN,
-        "Neb": DSOType.DN,
-        "NonEx": DSOType.OTHER,
-        "Nova": DSOType.OTHER,
-        "OCl": DSOType.OC,
-        "Other": DSOType.OTHER,
-        "PN": DSOType.PN,
-        "RfN": DSOType.RN,
-        "SNR": DSOType.SNR,
-    }
-    ngc_df[ObjectTableLabels.DSO_TYPE] = (
-        ngc_df["Type"].map(ngc_type_map.get).fillna(cast(DSOType, DSOType.OTHER))
-    )
-
-    # Standardize dimensions
-    ngc_df[ObjectTableLabels.SIZE_MAJOR] = pd.to_numeric(
-        ngc_df["MajAx"], errors="coerce"
-    )
-    ngc_df[ObjectTableLabels.SIZE_MINOR] = cast(
-        pd.Series, pd.to_numeric(ngc_df["MinAx"], errors="coerce")
-    ).fillna(ngc_df[ObjectTableLabels.SIZE_MAJOR])
-
-    # Drop redundant columns
-    ngc_df.drop(columns=["MajAx", "MinAx"], inplace=True)
-
-    # Map constellation abbreviations to full names
-    ngc_df["Constellation"] = ngc_df["Constellation"].map(constellation_map)  # type: ignore
-
-    # Set proper dtypes for string columns
-    string_columns = ["Name", "Type", "Constellation", "NGC"]
-    for column in string_columns:
-        if column in ngc_df.columns:
-            ngc_df[column] = ngc_df[column].astype("string")
-
-    # Vectorized RA and Dec parsing (Optimization: replaces slow row-wise apply)
-    # This provides a ~40% speedup for NGC loading by avoiding ~14k row-wise function calls.
-    ras_split = ngc_df["RA"].str.split(":", expand=True)
-    # Ensure 3 columns exist to avoid KeyErrors on partial data
-    for col in range(3):
-        if col not in ras_split.columns:
-            ras_split[col] = 0
-    h_ra = cast(pd.Series, pd.to_numeric(ras_split[0], errors="coerce"))
-    m_ra = cast(pd.Series, pd.to_numeric(ras_split[1], errors="coerce")).fillna(0)
-    s_ra = cast(pd.Series, pd.to_numeric(ras_split[2], errors="coerce")).fillna(0)
-    ra_hours = cast(pd.Series, h_ra + m_ra / 60.0 + s_ra / 3600.0)
-
-    decs_signs = cast(
-        pd.Series, ngc_df["Dec"].str.startswith("-", na=False).map({True: -1, False: 1})
-    )
-    decs_split = ngc_df["Dec"].str.lstrip("+-").str.split(":", expand=True)
-    for col in range(3):
-        if col not in decs_split.columns:
-            decs_split[col] = 0
-    h_dec = cast(pd.Series, pd.to_numeric(decs_split[0], errors="coerce"))
-    m_dec = cast(pd.Series, pd.to_numeric(decs_split[1], errors="coerce")).fillna(0)
-    s_dec = cast(pd.Series, pd.to_numeric(decs_split[2], errors="coerce")).fillna(0)
-    dec_degrees = cast(pd.Series, decs_signs * (h_dec + m_dec / 60.0 + s_dec / 3600.0))
-
-    # Store float versions for performance-critical filtering and calculations
-    # to avoid Pint and Skyfield object overhead in high-frequency loops.
-    # Optimization: We skip eager creation of skyfield_object, Magnitude (Quantity)
-    # and Size (Quantity) for all 14k entries to save ~1s on catalog load.
-    # These are restored lazily only for visible objects in NGC.get_visible().
-    ngc_df["ra_hours"] = ra_hours.values
-    ngc_df["dec_degrees"] = dec_degrees.values
-
-    magnitudes = cast(
-        pd.Series, pd.to_numeric(ngc_df["Magnitude"], errors="coerce")
-    ).fillna(99)
-    ngc_df["Magnitude_float"] = magnitudes.values
-
-    # Optimization: We keep 'Magnitude' and 'Size' as raw floats/strings in the catalog
-    # for performance. They are converted to Pint Quantities lazily in NGC.get_visible().
-    # We use object dtype to avoid FutureWarnings when restoring Quantities.
-    ngc_df["Magnitude"] = cast(pd.Series, magnitudes).astype(object)
-    ngc_df[ObjectTableLabels.SIZE_MAJOR] = ngc_df[ObjectTableLabels.SIZE_MAJOR].astype(
-        object
-    )
-    ngc_df[ObjectTableLabels.SIZE_MINOR] = ngc_df[ObjectTableLabels.SIZE_MINOR].astype(
-        object
-    )
-
-    # Add external links (vectorized list comprehension)
-    quoted_names = [urllib.parse.quote(str(x)) for x in ngc_df["Name"]]
-    ngc_df[ObjectTableLabels.SIMBAD] = [
-        f"https://simbad.u-strasbg.fr/simbad/sim-basic?Ident={q}" for q in quoted_names
-    ]
-    ngc_df[ObjectTableLabels.ALADIN] = [
-        f"https://aladin.cds.unistra.fr/AladinLite/?target={q}" for q in quoted_names
-    ]
-    ngc_df[ObjectTableLabels.ASTROBIN] = [
-        f"https://www.astrobin.com/search/?q={q}" for q in quoted_names
-    ]
-
-    return ngc_df
-
-
-def _load_bright_stars_with_units():
-    # Load bright stars catalogue data
-    bright_stars_df = pd.read_csv(
-        str(resources.files("apts").joinpath("data/bright_stars.csv"))
-    )
-
-    # Set proper dtypes for string columns
-    string_columns = ["Name"]
-    for column in string_columns:
-        bright_stars_df[column] = bright_stars_df[column].astype("string")
-
-    # Pre-calculate Skyfield objects using raw floats before wrapping in Quantities
-    # Optimization: avoiding row-wise apply and costly Quantity.to().magnitude calls
-    bright_stars_df["skyfield_object"] = [  # type: ignore
-        Star(ra_hours=ra, dec_degrees=dec)
-        for ra, dec in zip(bright_stars_df["RA"], bright_stars_df["Dec"])
-    ]
-
-    # Store float versions for performance-critical filtering and calculations
-    # to avoid Pint and Skyfield object overhead in high-frequency loops.
-    bright_stars_df["ra_hours"] = bright_stars_df["RA"].values
-    bright_stars_df["dec_degrees"] = bright_stars_df["Dec"].values
-    bright_stars_df["Magnitude_float"] = bright_stars_df["Magnitude"].values
-
-    # Convert columns to quantities with units (vectorized)
-    # Optimization: list(values * unit) is much faster than Series.apply()
-    ureg = get_unit_registry()
-    bright_stars_df["RA"] = list(bright_stars_df["RA"].values * ureg.hour)
-    bright_stars_df["Dec"] = list(bright_stars_df["Dec"].values * ureg.degree)
-    bright_stars_df["Magnitude"] = list(bright_stars_df["Magnitude"].values * ureg.mag)
-
-    return bright_stars_df
-
-
-class Catalogs:
-    """
-    This class provides access to astronomical catalogs.
-    To improve application startup time, catalogs are loaded lazily on first access.
-    """
-
-    def __init__(self):
-        pass
-
-    @property
-    def MESSIER(self) -> pd.DataFrame:
-        # Lazy load Messier catalog to avoid import overhead
-        global _messier_df
-        if _messier_df is None:
-            logger.info("Loading Messier catalog...")
-            _messier_df = _load_messier_with_units()
-        return _messier_df  # type: ignore
-
-    @property
-    def NGC(self) -> pd.DataFrame:
-        # Lazy load NGC catalog to avoid import overhead
-        global _ngc_df
-        if _ngc_df is None:
-            logger.info("Loading NGC catalog...")
-            _ngc_df = _load_ngc_with_units()
-        return _ngc_df  # type: ignore
-
-    @property
-    def BRIGHT_STARS(self) -> pd.DataFrame:
-        # Lazy load Bright Stars catalog to avoid import overhead
-        global _bright_stars_df
-        if _bright_stars_df is None:
-            logger.info("Loading Bright Stars catalog...")
-            _bright_stars_df = _load_bright_stars_with_units()
-        return _bright_stars_df  # type: ignore
-
-
-def initialize_catalogs():
-    logger.info("Initializing catalogs...")
-    catalogs = Catalogs()
-    _ = catalogs.MESSIER
-    _ = catalogs.NGC
-    _ = catalogs.BRIGHT_STARS
