+++ apts/catalogs/__init__.py
@@ -1,0 +1,3 @@
+from .base import Catalogs, initialize_catalogs
+
+__all__ = ["Catalogs", "initialize_catalogs"]
